+++ src/IMU.hpp
@@ -3,9 +3,9 @@
 
 class IMU {
  public:
-  virtual float get_yaw() = 0;
-  virtual float get_pitch() = 0;
-  virtual float get_roll() = 0;
+  virtual double get_yaw() = 0;
+  virtual double get_pitch() = 0;
+  virtual double get_roll() = 0;
 };
 
 #endif  // SRC_IMU_HPP_
